Implement idp attribute mapper with type support
This is a first draft implementation meant to spark a discussion about
the proper implementation. The current state does work but has some
bad workarounds specifically around setting the type in the saml
attribute. This probably should be refactored by changing the
`Attribute` itself to allow to provide a type and set it when creating
instead of having the string replacement.